Introduce a send_core_update_notification_email filter to control whether to send update notifications.
Merges [25882] to the 3.7 branch.

Currently, this is dependent on whether the automatic updater is enabled (generally). This is likely to be decoupled in a future release.

This refocuses the auto_core_update_send_email filter to specifically be for success/fail emails for auto updates, while the new filter is more general.

see #25603.

Built from https://develop.svn.wordpress.org/branches/3.7@25885


git-svn-id: http://core.svn.wordpress.org/branches/3.7@25797 1a063a9b-81f0-0310-95a4-ce76da25c4cd