When asserting microtime output as a number, make it a number
microtime is by default a string. Doing a greater then or less than check of that string is a bad idea since it uses the first part (the micro part of microtime) rather then the actual time. This adds a helper to convert microtime output into a float which we can then use to properly compare the output of microtime.

This fixes an intermittent test failure.

fixes #30336


Built from https://develop.svn.wordpress.org/trunk@30337


git-svn-id: http://core.svn.wordpress.org/trunk@30336 1a063a9b-81f0-0310-95a4-ce76da25c4cd